requirements.txt
asgiref==2.3.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.4.1
Automat==0.6.0
bleach==2.1.3
certifi==2018.4.16
channels==2.1.1
constantly==15.1.0
daphne==2.1.1
Django==2.0.4
django-carton==1.2.1
django-crispy-forms==1.7.2
django-registration-redux==2.4
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
squareconnect==2.6.1
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
webencodings==0.5.1
zope.interface==4.5.0
